adds web utils
